REFACTOR: extract menu routes api implementation to the controller
